requirements.txt
selenium>=4.0.0
pytest>=7.0.0
pytest-xdist
requests
Appium-Python-Client
allure-pytest
python-dotenv